Rename package curio_controller to ackermann_drive_controller
1. Rename the package curio_controller to ackermann_drive_controller and update affected packages.
