+++ ackermann_drive_controller/config/AckermannDriveController.cfg
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-PACKAGE = 'curio_controller'
+PACKAGE = 'ackermann_drive_controller'
 
 from dynamic_reconfigure.parameter_generator_catkin import ParameterGenerator, bool_t, double_t
 
@@ -15,4 +15,4 @@ gen.add("wheel_separation_multiplier", double_t, 0, "Wheel separation multiplier
 gen.add("publish_rate", double_t, 0, "Publish rate of odom.", 50.0, 0.0, 2000.0)
 gen.add("enable_odom_tf", bool_t, 0, "Publish odom frame to tf.", True)
 
-exit(gen.generate(PACKAGE, "curio_controller", "DiffDriveController"))
+exit(gen.generate(PACKAGE, "ackermann_drive_controller", "AckermannDriveController"))
