+++ ackermann_drive_controller/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="2">
-    <name>curio_controller</name>
+    <name>ackermann_drive_controller</name>
     <version>0.0.0</version>
-    <description>Controllers for Curio, a Sawppy Rover robot</description>
+    <description>Ackermann drive controller for a 6 wheel drive 4 wheel steering robot</description>
     <maintainer email="rhys.manwaring@me.com">Rhys Mainwaring</maintainer>
     <license>BSD-3-Clause</license>
     <author email="rhys.manwaring@me.com">Rhys Mainwaring</author>
@@ -21,6 +21,6 @@
     <depend>urdf</depend>
 
     <export>
-        <controller_interface plugin="${prefix}/curio_controller_plugins.xml" />
+        <controller_interface plugin="${prefix}/ackermann_drive_controller_plugins.xml" />
     </export>
 </package>
