--- curio_control/config/control_diff_drive.yaml
@@ -1,60 +1,0 @@
-# Configuration for the diff drive controller
-# 
-curio_controller:
-  # Resources
-  #   ROS diff_drive_controller wiki
-  #     http://wiki.ros.org/diff_drive_controller
-  # 
-  type: 'curio_controller/DiffDriveController'
-  publish_rate: 50
-
-  # Joints
-  left_wheel: ['front_left_wheel_joint', 'mid_left_wheel_joint', 'back_left_wheel_joint']
-  right_wheel: ['front_right_wheel_joint', 'mid_right_wheel_joint', 'back_right_wheel_joint']
-
-  # Geometry
-  wheel_radius: 0.060
-  wheel_separation: 0.52
-  wheel_radius_multiplier: 1.0        # default: 1.0
-  wheel_separation_multiplier: 1.0    # default 1.0
-
-  # Odometry covariances for the encoder output of the robot. These values should
-  # be tuned to your robot's sample odometry data, but these values are a good place
-  # to start
-  pose_covariance_diagonal: [0.001, 0.001, 0.001, 0.001, 0.001, 0.03]
-  twist_covariance_diagonal: [0.001, 0.001, 0.001, 0.001, 0.001, 0.03]
-
-  # Top level frame (link) of the robot description
-  base_frame_id: 'base_link'
-
-  # Transform from odom -> base_link: Enable if NOT using ekf_localization. 
-  enable_odom_tf: true
-  odom_frame_id: '/odom'
-
-  # Disable transform from odom -> base_link: Disable if USING ekf_localization. 
-  # 
-  # Also disable if we are using the Intel Realsense T265 tracking camera
-  # which provides a transform from odom to camera_pose_link.
-  # 
-  #   enable_odom_tf: false
-
-  # The RoboClaw motor controller provides velocity data. 
-  estimate_velocity_from_position: false
-
-  # Commands
-  publish_cmd: true
-  allow_multiple_cmd_vel_publishers: false
-
-  # Velocity and acceleration limits for the robot
-  linear:
-    x:
-      has_velocity_limits    : true
-      max_velocity           : 1.0   # m/s
-      has_acceleration_limits: true
-      max_acceleration       : 3.0   # m/s^2
-  angular:
-    z:
-      has_velocity_limits    : true
-      max_velocity           : 1.0   # rad/s
-      has_acceleration_limits: true
-      max_acceleration       : 3.0   # rad/s^2
