+++ curio_control/launch/control.launch
@@ -4,12 +4,12 @@
     </include>
 
     <!-- Load controller parameters -->
-    <!-- <rosparam command="load" file="$(find curio_control)/config/control_diff_drive.yaml" /> -->
+    <rosparam command="load" file="$(find curio_control)/config/control_ackermann_drive.yaml" />
     <rosparam command="load" file="$(find curio_control)/config/control_joint_state.yaml" />
 
     <!-- Spawn the controllers -->
     <!-- <node pkg="controller_manager" type="spawner" name="base_controller_spawner"
-        args="joint_state_publisher diff_drive_controller" /> -->
+        args="joint_state_publisher ackermann_drive_controller" /> -->
     <node pkg="controller_manager" type="spawner" name="base_controller_spawner"
         args="joint_state_publisher" />
 
@@ -33,7 +33,7 @@
     -->
     <!-- <node pkg="twist_mux" type="twist_mux" name="twist_mux">
         <rosparam command="load" file="$(find curio_control)/config/twist_mux.yaml" />
-        <remap from="cmd_vel_out" to="diff_drive_controller/cmd_vel"/>
+        <remap from="cmd_vel_out" to="ackermann_drive_controller/cmd_vel"/>
     </node> -->
 
 </launch>
