+++ curio_control/package.xml
@@ -10,9 +10,9 @@
   <buildtool_depend>catkin</buildtool_depend>
   <build_depend>roslaunch</build_depend>
 
+  <exec_depend>ackermann_drive_controller</exec_depend>
   <exec_depend>controller_manager</exec_depend>
   <exec_depend>curio_description</exec_depend>
-  <!-- <exec_depend>diff_drive_controller</exec_depend> -->
   <exec_depend>joint_state_controller</exec_depend>
   <exec_depend>robot_state_publisher</exec_depend>
   <!-- <exec_depend>twist_mux</exec_depend> -->
