--- curio_controller/curio_controller_plugins.xml
@@ -1,19 +1,0 @@
-<library path="lib/libcurio_controller">
-
-  <class name="curio_controller/DiffDriveController"
-    type="curio_controller::DiffDriveController"
-    base_class_type="controller_interface::ControllerBase">
-    <description>
-      The DiffDriveController tracks velocity commands. It expects 2 VelocityJointInterface type of hardware interfaces.
-    </description>
-  </class>
-
-  <class name="curio_controller/AckermannDriveController"
-    type="curio_controller::AckermannDriveController"
-    base_class_type="controller_interface::ControllerBase">
-    <description>
-      The DiffDriveController tracks velocity commands. It expects 2 VelocityJointInterface type of hardware interfaces.
-    </description>
-  </class>
-
-</library>
