--- curio_controller/include/curio_controller/ackermann_drive_controller.h
@@ -1,47 +1,0 @@
-//
-//  Software License Agreement (BSD-3-Clause)
-//   
-//  Copyright (c) 2019 Rhys Mainwaring
-//  All rights reserved
-//   
-//  Redistribution and use in source and binary forms, with or without
-//  modification, are permitted provided that the following conditions
-//  are met:
-//
-//  1.  Redistributions of source code must retain the above copyright
-//      notice, this list of conditions and the following disclaimer.
-//
-//  2.  Redistributions in binary form must reproduce the above
-//      copyright notice, this list of conditions and the following
-//      disclaimer in the documentation and/or other materials provided
-//      with the distribution.
-//
-//  3.  Neither the name of the copyright holder nor the names of its
-//      contributors may be used to endorse or promote products derived
-//      from this software without specific prior written permission.
-// 
-//  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-//  "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-//  LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
-//  FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
-//  COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
-//  INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
-//  BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-//  LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-//  CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
-//  LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
-//  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-//  POSSIBILITY OF SUCH DAMAGE.
-//
-
-#ifndef CURIO_CONTROLLER_ACKERMANN_DRIVE_CONTROLLER_H_
-#define CURIO_CONTROLLER_ACKERMANN_DRIVE_CONTROLLER_H_
-
-namespace curio_controller
-{
-
-
-
-} // namespace curio_controller
-
-#endif // CURIO_CONTROLLER_ACKERMANN_DRIVE_CONTROLLER_H_
