--- curio_controller/launch/control.launch
@@ -1,39 +1,0 @@
-<launch>
-    <!-- Load robot description -->
-    <include file="$(find curio_description)/launch/description.launch" >
-    </include>
-
-    <!-- Load controller parameters -->
-    <!-- <rosparam command="load" file="$(find curio_control)/config/control_diff_drive.yaml" /> -->
-    <rosparam command="load" file="$(find curio_control)/config/control_joint_state.yaml" />
-
-    <!-- Spawn the controllers -->
-    <!-- <node pkg="controller_manager" type="spawner" name="base_controller_spawner"
-        args="joint_state_publisher diff_drive_controller" /> -->
-    <node pkg="controller_manager" type="spawner" name="base_controller_spawner"
-        args="joint_state_publisher" />
-
-    <!-- EKF localization
-    
-        The robot_localization node also broadcasts the transform from:
-            odom -> base_link
-    -->    
-    <!-- <node pkg="robot_localization" type="ekf_localization_node" name="ekf_localization_local">
-        <rosparam command="load" file="$(find curio_control)/config/ekf_localization_local.yaml" />
-    </node> -->
-   
-    <!-- Launch the robot state publisher -->
-    <node name="robot_state_publisher" pkg="robot_state_publisher" type="state_publisher" />
-
-    <!-- Twist multiplexer
-    
-        Resources:
-            http://wiki.ros.org/twist_mux
-            https://github.com/ros-teleop/twist_mux
-    -->
-    <!-- <node pkg="twist_mux" type="twist_mux" name="twist_mux">
-        <rosparam command="load" file="$(find curio_control)/config/twist_mux.yaml" />
-        <remap from="cmd_vel_out" to="diff_drive_controller/cmd_vel"/>
-    </node> -->
-
-</launch>
