+++ curio_viz/package.xml
@@ -8,7 +8,7 @@
     <author email="rhys.mainwaring@me.com">Rhys Mainwaring</author>
     <buildtool_depend>catkin</buildtool_depend>
     <build_depend>roslaunch</build_depend>
-    <exec_depend>wild1_description</exec_depend>
+    <exec_depend>curio_description</exec_depend>
     <exec_depend>joint_state_publisher</exec_depend>
     <exec_depend>robot_state_publisher</exec_depend>
     <exec_depend>rviz</exec_depend>
